requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit_learn
seaborn 
catboost
xgboost
mlflow
#-e .